+++ src/generators/content/MultilineContentGenerator.ts
@@ -7,7 +7,7 @@ export class MultilineContentGenerator extends BaseContentGenerator {
     }
 
     public generate(vars: GeneratorVars): string {
-        if (this._source.values && this._source.values.length) {
+        if (this._source.values && this._source.values instanceof Array && this._source.values.length) {
             return this._source.values.join("\n");
         }
         return "";